detail rtspstream

+++ model/rtsp_resource.go
@@ -1,11 +1,11 @@
 package model
 
 type RTSPResource struct {
-	ID           int64  `gorm:"id"`
+	ID           int  `gorm:"id"`
 	URL          string `gorm:"url"`
 	Position     string `gorm:"position"`
-	AlgModel     int64  `gorm:"model"`
-	EncodeNeeded int8   `gorm:"encodeneeded"`
+	AlgModel     int  `gorm:"model"`
+	EncodeNeeded int   `gorm:"encodeneeded"`
 }
 
 func (r RTSPResource) TableName() string {
